requirements.txt
django
django-imagekit
django-crispy-forms
